+++ R/vcfR_conversion.R
@@ -30,6 +30,7 @@
 #' 
 #' @seealso
 #' \code{\link{extract.gt}},
+#' \code{\link{alleles2consensus}},
 #' \code{\link[adegenet]{df2genind}},
 #' \code{\link[adegenet]{genind}},
 #' \href{http://cran.r-project.org/package=pegas}{pegas},