+++ tests/testthat.R
@@ -24,6 +24,8 @@ test_file("testthat/test_vcfR_methods.R")
 test_file("testthat/test_addID.R")
 test_file("testthat/test_maf.R")
 
+test_file("testthat/test_genotype_matrix_functions.R")
+
 #test_file("testthat/test_rank_variants.R")
 #test_file("testthat/test_var_window.R")
 #test_file("testthat/test_windowing.R")