+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # vcfR 1.0.0.9000
 
+
+* `gt2alleles` handles missing data ('.')
 * `read.vcfR()` checks for and removes carriage returns (Windows)'
 * `vcfR2DNAbin` converts 'NA' to 'n' prior to conversion to DNAbin
 * `chromR2vcfR` implements use.mask